+++ lucee-java/lucee-core/src/lucee/runtime/tag/Http41.java
@@ -949,7 +949,7 @@ public final class Http41 extends BodyTagImpl implements Http {
     	// set timeout
 			if(this.timeout==null) { // not set
 				this.timeout=PageContextUtil.remainingTime(pageContext);
-				if(this.timeout.getSeconds()<=0)
+				if(((int)this.timeout.getSeconds())<=0)
 					throw CFMLFactoryImpl.createRequestTimeoutException(pageContext);
     		}
 			print.e(this.timeout);
